FIX CONNECT TO DB AND READ CSV YESS. TO SLOW we need to make it faster

+++ ABMB/Models/OldFlight.cs
@@ -6,18 +6,18 @@ public class OldFlight
     public int Year { get; set; }
     public int Month { get; set; }
     public int Day { get; set; }
-    public int DepTime { get; set; }
+    public double DepTime { get; set; } // Changed to double
     public int SchedDepTime { get; set; }
-    public int DepDelay { get; set; }
-    public int ArrTime { get; set; }
+    public double DepDelay { get; set; } // Changed to double
+    public double ArrTime { get; set; } // Changed to double
     public int SchedArrTime { get; set; }
-    public int ArrDelay { get; set; }
+    public double ArrDelay { get; set; } // Changed to double
     public string Carrier { get; set; }
     public int FlightNumber { get; set; }
     public string TailNum { get; set; }
     public string Origin { get; set; }
     public string Destination { get; set; }
-    public int AirTime { get; set; }
+    public double AirTime { get; set; } // Changed to double
     public int Distance { get; set; }
     public int Hour { get; set; }
     public int Minute { get; set; }
